refactor: factories, handlers and model structure
