+++ microservices/consumer-eph/src/main/java/com/microsoft/azure/eventhubs/samples/testconsumereph/Models/EventProcessor.java
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project root for full license information.
  */
-package com.microsoft.azure.eventhubs.samples.testconsumereph;
+package com.microsoft.azure.eventhubs.samples.testconsumereph.Models;
 
 import com.microsoft.azure.eventhubs.EventData;
 import com.microsoft.azure.eventprocessorhost.CloseReason;
